Introduce Tagaro::RenderBackend.
This replaces RendererModule and parts of Renderer. As a first step, the
RendererModule classes are unexported and removed from the public interface
of Renderer. Renderer is not ported to RenderBackend in this step; the port
will move the responsibility for pixmap fetching from Renderer to Sprite, and
as the last step eliminate the Renderer class.