Instantiate backends in Tagaro::Theme. Add infrastructure for sprite routing.
The Renderer class is ready to be merged with ThemeProvider.

Tagaro::Theme now has the capability to represent multiple backends.
Sprite keys can be mapped to different element keys in the backends, and
distributed among backends. This mechanism does not have a public interface
yet, so all sprite keys are mapped unchanged to the default backend by now.