Board/SpriteObjectItem: make relationship more robust
Tagaro::Board needs to know about any SpriteObjectItem instances inside
it. The old approach using the ItemChildAddedChange notification et.al.
is error-prone, complicated (due to half-finished constructors) and does
not find grandchildren.

The new approach places the duty of relating Board to SpriteObjectItem
on the SpriteObjectItem. The related ItemParentChange is more robust and
reliable. Especially, finding grandparents of a certain kind is easier
than finding grandchildren.